backend/requirements.txt
flask
flask-cors
flask_sqlalchemy
python-dotenv
pymysql
pyjwt
werkzeug
